+++ packages/terra-core-docs/src/terra-dev-site/doc/button/example/A11yIconButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import IconPrinter from 'terra-icon/lib/icon/IconPrinter';
-import { IconButton, IconTypes } from 'terra-button';
+import Button, { IconTypes } from 'terra-button';
 import Card from 'terra-card';
 import classNames from 'classnames/bind';
 import styles from './ButtonDocCommon.module.scss';
@@ -13,9 +13,9 @@ const DecorativeIcon = <IconPrinter />;
 const A11yIconButton = () => (
   <Card>
     <Card.Body>
-      <IconButton icon={Icon} isIconOnly iconType={IconTypes.INFORMATIVE} className={cx('button')} />
-      <IconButton icon={Icon} text="All Documents" iconType={IconTypes.INFORMATIVE} className={cx('button')} />
-      <IconButton icon={DecorativeIcon} iconType={IconTypes.INFORMATIVE} text="Print" className={cx('button')} />
+      <Button icon={Icon} isIconOnly iconType={IconTypes.INFORMATIVE} className={cx('button')} />
+      <Button icon={Icon} text="All Documents" iconType={IconTypes.INFORMATIVE} className={cx('button')} />
+      <Button icon={DecorativeIcon} iconType={IconTypes.DECORATIVE} text="Print" className={cx('button')} />
     </Card.Body>
   </Card>
 );