+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,12 +4,11 @@
 
 * Added 
   * Added Accessibility Guide for `terra-toggle`.
+  * Added A11y guide and examples for `terra-button`.
 
 * Updated
+  * Updated imports of decorative icons.
   * Updated jest & wdio tests.
-
-* Added
-  * Added A11y guide and examples for `terra-button`.
 
 ## 1.20.0 - (February 7, 2023)